requirements.txt
asgiref==3.6.0
astroid==2.13.2
defusedxml==0.7.1
diff-match-patch==20200713
dill==0.3.6
Django==4.1.5
django-import-export==3.0.2
djangorestframework==3.14.0
et-xmlfile==1.1.0
isort==5.11.4
lazy-object-proxy==1.9.0
MarkupPy==1.14
mccabe==0.7.0
odfpy==1.4.1
openpyxl==3.0.10
platformdirs==2.6.2
pylint==2.15.10
pylint-plugin-utils==0.7
pytz==2022.7.1
PyYAML==6.0
sqlparse==0.4.3
tablib==3.3.0
tomli==2.0.1
tomlkit==0.11.6
typing_extensions==4.4.0
wrapt==1.14.1
xlrd==2.0.1
xlwt==1.3.0
